+++ frontend/src/utils.js
@@ -5,6 +5,8 @@ export const VerifiableType = {
     PRESENTATION: 'VerifiablePresentation'
 };
 
+const IPFS_GATEWAYS = ['ipfs.io', 'ipfs.ssi.eecc.de']
+
 export function isURL(url) {
     if (typeof url != 'string') return false;
     return url.startsWith('https://');
@@ -35,7 +37,48 @@ export function getHolder(presentation) {
     return proof.verificationMethod.split('#')[0];
 }
 
+export async function fetchIPFS(IPFSUrl) {
+
+    var document;
+
+    await Promise.any(IPFS_GATEWAYS.map(async (gateway) => {
+
+        return await fetch(`https://${gateway}/ipfs/${IPFSUrl.split('ipfs://')[1]}`);
+
+    }))
+        .then((result) => {
+
+            document = result;
+
+        })
+        .catch((error) => {
+            console.error(error)
+        })
+
+    if (!document) throw Error('Fetching from IPFS failed');
+
+    return document;
+
+}
+
+
+const documentLoader = async (url) => {
+    let document;
+    if (url.startsWith('ipfs://')) {
+
+        document = await fetchIPFS(url)
+
+    } else document = await fetch(url);
+
+    return {
+        contextUrl: null,
+        document: await document.json(),
+        documentUrl: url
+    };
+};
+
+
 export async function getContext(credential) {
-    const resolved = await jsonld.processContext(await jsonld.processContext(null, null), credential);
+    const resolved = await jsonld.processContext(await jsonld.processContext(null, null), credential, { documentLoader });
     return resolved.mappings;
 }